With AppContext; custom queries work now
