+++ main/src/components/InteractionLayout/InteractionLayout.jsx
@@ -1,6 +1,6 @@
+import { Component } from "react";
 import { Layout } from "react-admin";
 import NavigationBar from "./NavigationBar/NavigationBar";
-import { Component } from "react";
 import SelectionMenu from "./SelectionMenu/SelectionMenu";
 
 // LOG let interactionLayoutCounter = 0;
@@ -12,6 +12,7 @@ import SelectionMenu from "./SelectionMenu/SelectionMenu";
  */
 function InteractionLayout(props) {
   // LOG console.log(`--- InteractionLayout #${++interactionLayoutCounter}`);
+  
   return <Layout {...props} appBar={NavigationBar} menu={SelectionMenu} />;
 }
